feat: HardHat not starts Coreminer as subprocess and can cummunicate with it
