+++ tui/app.py
@@ -1,7 +1,7 @@
 from textual.app import App
-from .views.grid_view import GridView
+from .views.main_view import MainView
 
 class HardHat(App):
     """Main Textual application."""
     def on_mount(self):
-        self.push_screen(GridView())
+        self.push_screen(MainView())
